requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
anyio==3.6.2
arandr==0.1.11
async-timeout==4.0.2
attrs==22.2.0
autocommand==2.2.2
Babel==2.9.1
beautifulsoup4==4.11.2
bs4==0.0.1
btrfsutil==6.1.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.1
Cython==0.29.33
exceptiongroup==1.1.1
fastapi==0.95.0
fastjsonschema==2.16.2
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
inflect==6.0.2
iniconfig==2.0.0
jaraco.context==4.3.0
jaraco.functools==3.5.2
jaraco.text==3.11.1
kazam==1.4.5
libtorrent==2.0.8
lxml==4.9.2
magic-filter==1.0.9
more-itertools==9.0.0
multidict==6.0.4
ordered-set==4.1.0
packaging==23.0
platformdirs==2.6.2
pluggy==1.0.0
ply==3.11
py3status==3.48
pycairo==1.23.0
pycparser==2.21
pydantic==1.10.4
PyGObject==3.42.2
pytest==7.2.2
pytest-asyncio==0.21.0
pytest-check==2.1.4
python-distutils-extra==2.39
python-dotenv==1.0.0
pytz==2022.7.1
pyxdg==0.28
ranger-fm==1.9.3
redis==4.5.3
requests==2.28.2
shiboken2==5.15.8
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.7
starlette==0.26.1
team==1.0
tomli==2.0.1
trove-classifiers==2023.2.9
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
validate-pyproject==0.12.1
XlsxWriter==3.0.9
yarl==1.8.2
